delete from fe

+++ MicdUp/server/database/publicSchema/mutations/recording.js
@@ -247,9 +247,9 @@ const likePost = {
 const deletePost = {
   type: MessageType,
   args: {
-    postId: {type: GraphQLID},
+    postId: { type: GraphQLID },
   },
-  async resolve(parent, {postId}, context) {
+  async resolve(parent, { postId }, context) {
     const session = await mongoose.startSession();
     session.startTransaction();
     returnObject = {
@@ -261,11 +261,14 @@ const deletePost = {
       const post = await Post.findOne({
         _id: postId,
       });
+      if (!context.profile || post.owner.toString() !== context.profile.id) {
+        throw new Error("You are not authorized to delete this post");
+      }
       if (!post) {
         throw new Error("post not found");
       }
       for (let i = 0; i < post.comments.length; i++) {
-        await Comment.findByIdAndDelete({ _id: post.comments[i]})
+        await Comment.findByIdAndDelete({ _id: post.comments[i] });
       }
       await post.save({ session });
       await Post.findByIdAndDelete(postId);
@@ -277,10 +280,10 @@ const deletePost = {
     } finally {
       session.endSession();
     }
-    console.log("Delete post successful.")
+    console.log("Delete post successful.");
     return returnObject;
-  }
-}
+  },
+};
 
 const commentToPost = {
   type: CommentType,
